requirements.txt
requests
python-dotenv
apache-airflow-providers-microsoft-azure
azure-storage-file-datalake
dbt-postgres